+++ src/euroeval/benchmark_modules/vllm.py
@@ -43,6 +43,7 @@ from ..enums import (
 from ..exceptions import (
     InvalidBenchmark,
     InvalidModel,
+    InvalidTask,
     NeedsEnvironmentVariable,
     NeedsExtraInstalled,
     NeedsSystemDependency,
@@ -62,6 +63,7 @@ from ..task_group_utils import (
     text_to_text,
     token_classification,
 )
+from ..tasks import NER
 from ..tokenisation_utils import (
     apply_chat_template,
     get_bos_token,
@@ -428,6 +430,9 @@ class VLLMModel(HuggingFaceEncoderModel):
             InvalidBenchmark:
                 If the dataset requires logprobs, but we could not get the first token
                 of each label in the dataset.
+            InvalidTask:
+                If the task requires structured output, but either is not NER or
+                does not define an output structure.
         """
         # Get stopping tokens
         stop_tokens: list[str] = self.custom_stop_tokens.copy()
@@ -487,21 +492,36 @@ class VLLMModel(HuggingFaceEncoderModel):
                 level=logging.DEBUG,
             )
         elif self.dataset_config.task.uses_structured_output:
-            ner_tag_names = list(self.dataset_config.prompt_label_mapping.values())
-            keys_and_their_types: dict[str, t.Any] = {
-                tag_name: (conlist(str, max_length=5), ...)
-                for tag_name in ner_tag_names
-            }
-            answer_format_class = create_model("AnswerFormat", **keys_and_their_types)
-            structured_generation_schema = answer_format_class.model_json_schema()
-            log_once(
-                "Using structured generation with the JSON schema: "
-                f"{json.dumps(structured_generation_schema, ensure_ascii=False)}",
-                level=logging.DEBUG,
-            )
-            structured_outputs = StructuredOutputsParams(
-                json=structured_generation_schema
-            )
+            if self.dataset_config.task.structured_output_format is not None:
+                structured_outputs = StructuredOutputsParams(
+                    json=self.dataset_config.task.structured_output_format.model_json_schema()
+                )
+            elif self.dataset_config.task == NER:
+                ner_tag_names = list(self.dataset_config.prompt_label_mapping.values())
+                keys_and_their_types: dict[str, t.Any] = {
+                    tag_name: (conlist(str, max_length=5), ...)
+                    for tag_name in ner_tag_names
+                }
+                answer_format_class = create_model(
+                    "AnswerFormat", **keys_and_their_types
+                )
+                structured_generation_schema = answer_format_class.model_json_schema()
+                log_once(
+                    "Using structured generation with the JSON schema: "
+                    f"{json.dumps(structured_generation_schema, ensure_ascii=False)}",
+                    level=logging.DEBUG,
+                )
+                structured_outputs = StructuredOutputsParams(
+                    json=structured_generation_schema
+                )
+            else:
+                raise InvalidTask(
+                    message=(
+                        "Task set to use structured out, but neither is an NER task "
+                        "nor defines an output structure "
+                        "- at least one of these must be true."
+                    )
+                )
         elif (
             self.dataset_config.task.uses_logprobs
             and self.dataset_config.labels