+++ src/euroeval/metrics/__init__.py
@@ -6,3 +6,4 @@ from .ifeval import *  # noqa: F403
 from .llm_as_a_judge import *  # noqa: F403
 from .pipeline import *  # noqa: F403
 from .speed import *  # noqa: F403
+from .tool_calling import *  # noqa: F403