+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
+- A task _Tool Calling_ and a dataset under this task _bfcl-v2 a subset of the
+  Berkeley Function Calling Leaderboard benchmark (v2).
+  Currently only supported for English.
 - Added the new Danish linguistic acceptability dataset DaLA. It's marked as
   unofficial for now. This was added by @N-essuno ✨
 - Added `--max-context-length` and `--vocabulary-size` CLI options (and corresponding